Fix handling of null in DecodeMapInterfaceType()
DecodeMapInterfaceType() is subtly broken for nil values, when given the
zero Value reflect.Value.SetMapIndex() does not add the element to the
map but instead delete it from the map, which means that the job spec

    job "docs" {
      region = "us"
      datacenters = ["us-west-1", "us-east-1"]
      type = "service"

      group "webs" {
          task "frontend" {
            driver = "docker"

            config {
              image = "hashicorp/web-frontend"
              auth  = null
            }
          }
      }
    }

does not roundtrip properly, the 'auth' attribute will be missing.
Avoiding the use of reflect.Value.SetMapIndex() by accessing directly
the underlying value fixes this issue.